tests/wpt/web-platform-tests/html/user-activation/activation-api-iframe-no-activate.tentative.html

<!DOCTYPE html>
<!--
   Tentative due to:
   https://github.com/whatwg/html/issues/1983
-->
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
  <h1>Clicking in parent frame does not propagate state to child</h1>
  <ol id="instructions">
    <li>Click this instruction text.
  </ol>
  <iframe id="child" width="200" height="200"
          src="resources/child-one.html">
  </iframe>
  <script>
    async_test(function(t) {
      var child = document.getElementById("child");
      assert_false(navigator.userActivation.isActive);
      assert_false(navigator.userActivation.hasBeenActive);
      window.addEventListener("message", t.step_func(event => {
        var msg = JSON.parse(event.data);
        if (msg.type == 'child-one-loaded') {
          // Child's states should be false after load.
          assert_false(msg.isActive);
          assert_false(msg.hasBeenActive);

          // Click in parent document.
          test_driver.click(document.getElementById('instructions'));
        } else if (msg.type == 'child-one-report') {
          // Only the transient state in child should be false after consumption below.
          assert_false(msg.isActive);
          assert_true(msg.hasBeenActive);
          t.done();
        }
      }));
      window.addEventListener("click", t.step_func(event => {
        assert_true(navigator.userActivation.isActive);
        assert_true(navigator.userActivation.hasBeenActive);

        // Opening a window should consume the activation.
        var win = window.open('404.html');
        win.close();
        assert_false(navigator.userActivation.isActive);
        assert_true(navigator.userActivation.hasBeenActive);

        // Ask child to report its state.
        child.contentWindow.postMessage(JSON.stringify({"type": "report"}), '*');
      }));
    }, "Values adjust on activity");
  </script>
</body>
</html>
